separate json_path related functions
