+++ src/compat/db_json.hpp
@@ -47,7 +47,7 @@ typedef void JSON_ITERATOR;
 /*
  * these also double as type precedence
  * INT and DOUBLE actually have the same precedence
-*/
+ */
 enum DB_JSON_TYPE
 {
   DB_JSON_NULL = 0,
@@ -61,7 +61,6 @@ enum DB_JSON_TYPE
   DB_JSON_BOOL,
 };
 
-/* C functions */
 bool db_json_is_valid (const char *json_str);
 const char *db_json_get_type_as_str (const JSON_DOC *document);
 unsigned int db_json_get_length (const JSON_DOC *document);
@@ -100,8 +99,6 @@ int db_json_keys_func (const JSON_DOC &doc, JSON_DOC &result_json, const char *r
 int db_json_array_append_func (const JSON_DOC *value, JSON_DOC &doc, const char *raw_path);
 int db_json_array_insert_func (const JSON_DOC *value, JSON_DOC &doc, const char *raw_path);
 int db_json_remove_func (JSON_DOC &doc, const char *raw_path);
-int db_json_paths_to_regex (const std::vector<std::string> &paths, std::vector<std::string> &regs,
-			    bool match_exactly = false);
 int db_json_search_func (JSON_DOC &doc, const DB_VALUE *pattern, const DB_VALUE *esc_char,
 			 std::vector<std::string> &paths, const std::vector<std::string> &starting_paths, bool find_all);
 int db_json_merge_patch_func (const JSON_DOC *source, JSON_DOC *&dest);
@@ -123,7 +120,6 @@ void db_json_delete_doc (JSON_DOC *&doc);
 void db_json_delete_validator (JSON_VALIDATOR *&validator);
 int db_json_validate_doc (JSON_VALIDATOR *validator, JSON_DOC *doc);
 bool db_json_are_validators_equal (JSON_VALIDATOR *val1, JSON_VALIDATOR *val2);
-void db_json_path_unquote_object_keys (std::string &sql_path);
 bool db_json_path_contains_wildcard (const char *sql_path);
 
 void db_json_iterator_next (JSON_ITERATOR &json_itr);
@@ -136,11 +132,7 @@ JSON_ITERATOR *db_json_create_iterator (const DB_JSON_TYPE &type);
 void db_json_delete_json_iterator (JSON_ITERATOR *&json_itr);
 void db_json_clear_json_iterator (JSON_ITERATOR *&json_itr);
 
-int db_json_path_validate_external (const char *pointer_path, bool allow_wildcards);
-int db_json_normalize_path (const char *pointer_path, JSON_PATH &json_path,
-			    bool allow_wildcards = true);
 DB_JSON_TYPE db_json_get_type (const JSON_DOC *doc);
-
 int db_json_get_int_from_document (const JSON_DOC *doc);
 std::int64_t db_json_get_bigint_from_document (const JSON_DOC *doc);
 double db_json_get_double_from_document (const JSON_DOC *doc);
@@ -164,7 +156,9 @@ bool db_json_doc_is_uncomparable (const JSON_DOC *doc);
 int db_value_to_json_doc (const DB_VALUE &db_val, REFPTR (JSON_DOC, json_doc));
 int db_value_to_json_value (const DB_VALUE &db_val, REFPTR (JSON_DOC, json_val));
 int db_value_to_json_path (const DB_VALUE *path_value, FUNC_TYPE fcode, const char **path_str);
-/* end of C functions */
+
+int db_json_path_validate_external (const char *pointer_path, bool allow_wildcards);
+void db_json_path_unquote_object_keys_external (std::string &sql_path);
 
 template <typename Fn, typename... Args>
 inline int
