+++ src/compat/db_json_types_internal.cpp
@@ -1,0 +1,99 @@
+#include "db_json_types_internal.hpp"
+
+/************************************************************************/
+/* JSON_DOC implementation                                              */
+/************************************************************************/
+
+bool JSON_DOC::IsLeaf()
+{
+  return !IsArray () && !IsObject ();
+}
+
+void *
+JSON_PRIVATE_ALLOCATOR::Malloc (size_t size)
+{
+  if (size)			//  behavior of malloc(0) is implementation defined.
+    {
+      char *p = (char *) db_private_alloc (NULL, size);
+      if (prm_get_bool_value (PRM_ID_JSON_LOG_ALLOCATIONS))
+	{
+	  er_print_callstack (ARG_FILE_LINE, "JSON_ALLOC: Traced pointer=%p\n", p);
+	}
+      return p;
+    }
+  else
+    {
+      return NULL;		// standardize to returning NULL.
+    }
+}
+
+void *
+JSON_PRIVATE_ALLOCATOR::Realloc (void *originalPtr, size_t originalSize, size_t newSize)
+{
+  (void) originalSize;
+  char *p;
+  if (newSize == 0)
+    {
+      db_private_free (NULL, originalPtr);
+      return NULL;
+    }
+  p = (char *) db_private_realloc (NULL, originalPtr, newSize);
+  if (prm_get_bool_value (PRM_ID_JSON_LOG_ALLOCATIONS))
+    {
+      er_print_callstack (ARG_FILE_LINE, "Traced pointer=%p\n", p);
+    }
+  return p;
+}
+
+void
+JSON_PRIVATE_ALLOCATOR::Free (void *ptr)
+{
+  db_private_free (NULL, ptr);
+}
+
+const bool JSON_PRIVATE_ALLOCATOR::kNeedFree = true;
+
+
+DB_JSON_TYPE
+db_json_get_type_of_value (const JSON_VALUE *val)
+{
+  if (val == NULL)
+    {
+      return DB_JSON_NULL;
+    }
+
+  if (val->IsString ())
+    {
+      return DB_JSON_STRING;
+    }
+  else if (val->IsInt ())
+    {
+      return DB_JSON_INT;
+    }
+  else if (val->IsInt64 ())
+    {
+      return DB_JSON_BIGINT;
+    }
+  else if (val->IsFloat () || val->IsDouble ())
+    {
+      return DB_JSON_DOUBLE;
+    }
+  else if (val->IsObject ())
+    {
+      return DB_JSON_OBJECT;
+    }
+  else if (val->IsArray ())
+    {
+      return DB_JSON_ARRAY;
+    }
+  else if (val->IsNull ())
+    {
+      return DB_JSON_NULL;
+    }
+  else if (val->IsBool ())
+    {
+      return DB_JSON_BOOL;
+    }
+
+  return DB_JSON_UNKNOWN;
+}
