+++ src/query/arithmetic.c
@@ -6343,7 +6343,7 @@ db_evaluate_json_search (DB_VALUE *result, DB_VALUE * const * args, const int nu
       char *escaped;
       size_t escaped_size;
 
-      db_json_path_unquote_object_keys (paths[0]);
+      db_json_path_unquote_object_keys_external (paths[0]);
       error_code = db_string_escape (paths[0].c_str(), paths[0].size (), &escaped, &escaped_size);
       if (error_code)
 	{
@@ -6366,7 +6366,7 @@ db_evaluate_json_search (DB_VALUE *result, DB_VALUE * const * args, const int nu
       char *escaped;
       size_t escaped_size;
 
-      db_json_path_unquote_object_keys (paths[i]);
+      db_json_path_unquote_object_keys_external (paths[i]);
       error_code = db_string_escape (paths[i].c_str (), paths[i].size (), &escaped, &escaped_size);
       if (error_code)
 	{
